+++ packages/elements/src/components.d.ts
@@ -5,7 +5,7 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { Alignment, ButtonType, ChipSurface, ColorScheme, DialogCloseAction, HorizontalLocation, ImageDecodingTypes, InputType, Locations, NavDrawerAnchor, NavDrawerVariant, SnackbarType, SpinnerType, TooltipTrigger, UserInputInterceptor, VerticalLocation, ViewModeUnion } from "./components/types";
+import { Alignment, ButtonType, ChipSurface, DialogCloseAction, HorizontalLocation, ImageDecodingTypes, InputType, KeyValue, Locations, NavDrawerAnchor, NavDrawerVariant, SnackbarType, SpinnerType, TooltipTrigger, UserInputInterceptor, VerticalLocation, ViewModeUnion } from "./components/types";
 import { Variants } from "./components/ino-button/ino-button";
 import { PickerTypeKeys } from "./components/ino-datepicker/picker-factory";
 import { Placement, Props } from "tippy.js";
@@ -13,17 +13,21 @@ import { SortDirection, SortDirectionChangeDetails } from "./interface";
 export namespace Components {
     interface InoAutocomplete {
         /**
-          * Timeout of the debouncing mechanism used when filtering the options.
+          * Number of ms the search function should be delayed after the user typed something.
          */
-        "debounceTimeout": number;
+        "debounce": number;
         /**
-          * Text to display when there are no options found.
+          * Text to display when there are no options found, where `$` is the placeholder for the input of the user.
          */
         "noOptionsText": string;
         /**
-          * Value of the autocomplete
+          * All options either as a string array or as an array of `{key: string; value: string}` objects.
          */
-        "value": any;
+        "options": string[] | KeyValue[];
+        /**
+          * The selected value.
+         */
+        "value": string | KeyValue | null;
     }
     interface InoButton {
         /**
@@ -136,10 +140,6 @@ export namespace Components {
         "value"?: string;
     }
     interface InoChip {
-        /**
-          * The name of the color scheme which is used to style the background and outline of this component.
-         */
-        "colorScheme": ColorScheme | 'default';
         /**
           * Disables all interactions.
          */
@@ -419,16 +419,6 @@ export namespace Components {
          */
         "topBottomLocation": VerticalLocation;
     }
-    interface InoFormRow {
-        /**
-          * The label for this form row which describes the form element.
-         */
-        "label"?: string;
-        /**
-          * An indicator which marks the contents of the form row as mandatory. If you use this make sure you also check for the values in your application logic.
-         */
-        "mandatory"?: boolean;
-    }
     interface InoHeader {
         /**
           * [DEPRECATED] Please use the default slot instead of this prop.
@@ -467,10 +457,6 @@ export namespace Components {
           * Sets the autofocus for this element.
          */
         "autoFocus"?: boolean;
-        /**
-          * The name of the color scheme which is used to style the background and outline of this component. Possible values: `primary` (default),  `secondary`, `success`, `warning`, `error`, `light`, `dark`.
-         */
-        "colorScheme"?: ColorScheme;
         /**
           * Disables this element.
          */
@@ -870,9 +856,9 @@ export namespace Components {
          */
         "attachToBody": boolean;
         /**
-          * Sets the color scheme of the popup Valid options include: 'primary', 'secondary', 'light', 'transparent'
+          * Sets the color scheme of the popup Valid options include: 'primary', 'transparent'
          */
-        "colorScheme": 'primary' | 'secondary' | 'light' | 'transparent';
+        "colorScheme": 'primary' | 'transparent';
         /**
           * Used to indicate if the popover should be controlled by itself (`false`) or manually by the `visible` property (`true`)
          */
@@ -965,10 +951,6 @@ export namespace Components {
         "value"?: string | number | null;
     }
     interface InoRange {
-        /**
-          * The name of the color scheme of this component. Possible values: `primary` (default), `secondary`, `success`, `warning`, `error`, `light`, `dark`.
-         */
-        "colorScheme"?: ColorScheme;
         /**
           * Disables this element.
          */
@@ -1097,20 +1079,6 @@ export namespace Components {
          */
         "value"?: string;
     }
-    interface InoSidebar {
-        /**
-          * Aligns the sidebar to the right (true) or left (false) side
-         */
-        "alignRight": boolean;
-        /**
-          * Name of the component
-         */
-        "name"?: string;
-        /**
-          * Expands the sidebar
-         */
-        "open": boolean;
-    }
     interface InoSnackbar {
         /**
           * The text to display for the action button. If no text is defined, the snack bar is displayed in an alternative feedback style.
@@ -1136,10 +1104,6 @@ export namespace Components {
     }
     interface InoSpinner {
         /**
-          * The name of the color scheme which is used to style the background and outline of this component. Possible values: `primary` (default),  `secondary`, `success`, `warning`, `error`, `light`, `dark`.
-         */
-        "colorScheme": ColorScheme;
-        /**
           * The height of this spinner (default = 40)
          */
         "height": number;
@@ -1161,10 +1125,6 @@ export namespace Components {
           * Marks this element as checked. (**unmanaged**)
          */
         "checked"?: boolean;
-        /**
-          * The name of the color scheme which is used to style the background and outline of this component. Possible values: `primary` (default),  `secondary`, `success`, `warning`, `error`, `light`, `dark`.
-         */
-        "colorScheme": ColorScheme;
         /**
           * Disables this element.
          */
@@ -1345,9 +1305,9 @@ export namespace Components {
     }
     interface InoTooltip {
         /**
-          * Sets the color scheme of the tooltip Valid options include: `primary`, `secondary` `light`, `transparent`
+          * Sets the color scheme of the tooltip Valid options include: `primary`, `transparent`
          */
-        "colorScheme": 'primary' | 'secondary' | 'light' | 'transparent';
+        "colorScheme": 'primary' | 'transparent';
         /**
           * The target id the tooltip belongs to. If not given, the tooltip is attached to the parent component.
          */
@@ -1455,13 +1415,13 @@ export interface InoSegmentButtonCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLInoSegmentButtonElement;
 }
+export interface InoSegmentGroupCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLInoSegmentGroupElement;
+}
 export interface InoSelectCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLInoSelectElement;
-}
-export interface InoSidebarCustomEvent<T> extends CustomEvent<T> {
-    detail: T;
-    target: HTMLInoSidebarElement;
 }
 export interface InoSnackbarCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -1569,12 +1529,6 @@ declare global {
     var HTMLInoFabSetElement: {
         prototype: HTMLInoFabSetElement;
         new (): HTMLInoFabSetElement;
-    };
-    interface HTMLInoFormRowElement extends Components.InoFormRow, HTMLStencilElement {
-    }
-    var HTMLInoFormRowElement: {
-        prototype: HTMLInoFormRowElement;
-        new (): HTMLInoFormRowElement;
     };
     interface HTMLInoHeaderElement extends Components.InoHeader, HTMLStencilElement {
     }
@@ -1726,12 +1680,6 @@ declare global {
         prototype: HTMLInoSelectElement;
         new (): HTMLInoSelectElement;
     };
-    interface HTMLInoSidebarElement extends Components.InoSidebar, HTMLStencilElement {
-    }
-    var HTMLInoSidebarElement: {
-        prototype: HTMLInoSidebarElement;
-        new (): HTMLInoSidebarElement;
-    };
     interface HTMLInoSnackbarElement extends Components.InoSnackbar, HTMLStencilElement {
     }
     var HTMLInoSnackbarElement: {
@@ -1800,7 +1748,6 @@ declare global {
         "ino-dialog": HTMLInoDialogElement;
         "ino-fab": HTMLInoFabElement;
         "ino-fab-set": HTMLInoFabSetElement;
-        "ino-form-row": HTMLInoFormRowElement;
         "ino-header": HTMLInoHeaderElement;
         "ino-icon": HTMLInoIconElement;
         "ino-icon-button": HTMLInoIconButtonElement;
@@ -1826,7 +1773,6 @@ declare global {
         "ino-segment-button": HTMLInoSegmentButtonElement;
         "ino-segment-group": HTMLInoSegmentGroupElement;
         "ino-select": HTMLInoSelectElement;
-        "ino-sidebar": HTMLInoSidebarElement;
         "ino-snackbar": HTMLInoSnackbarElement;
         "ino-spinner": HTMLInoSpinnerElement;
         "ino-switch": HTMLInoSwitchElement;
@@ -1841,21 +1787,25 @@ declare global {
 declare namespace LocalJSX {
     interface InoAutocomplete {
         /**
-          * Timeout of the debouncing mechanism used when filtering the options.
+          * Number of ms the search function should be delayed after the user typed something.
          */
-        "debounceTimeout"?: number;
+        "debounce"?: number;
         /**
-          * Text to display when there are no options found.
+          * Text to display when there are no options found, where `$` is the placeholder for the input of the user.
          */
         "noOptionsText"?: string;
         /**
-          * Emits in three ways:  1. Clicking on an option 2. Pressing `Enter` while an option is selected 3. Entering a valid value and blurring the input element  Contains one of the texts provided by the `<ino-options>`s.
+          * Emits the list item the user clicked on either as a string or a `{key: string; value: string}` object depending on the provided options.  Trigger on two occasions: 1. The user clicked on a list-item. 2. The user types in a string that matches an option and blurs the input
          */
-        "onValueChange"?: (event: InoAutocompleteCustomEvent<string | null>) => void;
+        "onValueChange"?: (event: InoAutocompleteCustomEvent<string | { key: string; value: string }>) => void;
         /**
-          * Value of the autocomplete
+          * All options either as a string array or as an array of `{key: string; value: string}` objects.
          */
-        "value"?: any;
+        "options": string[] | KeyValue[];
+        /**
+          * The selected value.
+         */
+        "value"?: string | KeyValue | null;
     }
     interface InoButton {
         /**
@@ -1976,10 +1926,6 @@ declare namespace LocalJSX {
         "value"?: string;
     }
     interface InoChip {
-        /**
-          * The name of the color scheme which is used to style the background and outline of this component.
-         */
-        "colorScheme"?: ColorScheme | 'default';
         /**
           * Disables all interactions.
          */
@@ -2271,16 +2217,6 @@ declare namespace LocalJSX {
          */
         "topBottomLocation"?: VerticalLocation;
     }
-    interface InoFormRow {
-        /**
-          * The label for this form row which describes the form element.
-         */
-        "label"?: string;
-        /**
-          * An indicator which marks the contents of the form row as mandatory. If you use this make sure you also check for the values in your application logic.
-         */
-        "mandatory"?: boolean;
-    }
     interface InoHeader {
         /**
           * [DEPRECATED] Please use the default slot instead of this prop.
@@ -2323,10 +2259,6 @@ declare namespace LocalJSX {
           * Sets the autofocus for this element.
          */
         "autoFocus"?: boolean;
-        /**
-          * The name of the color scheme which is used to style the background and outline of this component. Possible values: `primary` (default),  `secondary`, `success`, `warning`, `error`, `light`, `dark`.
-         */
-        "colorScheme"?: ColorScheme;
         /**
           * Disables this element.
          */
@@ -2757,9 +2689,9 @@ declare namespace LocalJSX {
          */
         "attachToBody"?: boolean;
         /**
-          * Sets the color scheme of the popup Valid options include: 'primary', 'secondary', 'light', 'transparent'
+          * Sets the color scheme of the popup Valid options include: 'primary', 'transparent'
          */
-        "colorScheme"?: 'primary' | 'secondary' | 'light' | 'transparent';
+        "colorScheme"?: 'primary' | 'transparent';
         /**
           * Used to indicate if the popover should be controlled by itself (`false`) or manually by the `visible` property (`true`)
          */
@@ -2861,10 +2793,6 @@ declare namespace LocalJSX {
     }
     interface InoRange {
         /**
-          * The name of the color scheme of this component. Possible values: `primary` (default), `secondary`, `success`, `warning`, `error`, `light`, `dark`.
-         */
-        "colorScheme"?: ColorScheme;
-        /**
           * Disables this element.
          */
         "disabled"?: boolean;
@@ -2953,6 +2881,10 @@ declare namespace LocalJSX {
          */
         "name"?: string;
         /**
+          * Forwards the `checkedChange` events of the `<ino-segment-buttons>` with its value as the detail.
+         */
+        "onValueChange"?: (event: InoSegmentGroupCustomEvent<any>) => void;
+        /**
           * Value of the segment group
          */
         "value"?: any;
@@ -3007,24 +2939,6 @@ declare namespace LocalJSX {
          */
         "value"?: string;
     }
-    interface InoSidebar {
-        /**
-          * Aligns the sidebar to the right (true) or left (false) side
-         */
-        "alignRight"?: boolean;
-        /**
-          * Name of the component
-         */
-        "name"?: string;
-        /**
-          * Emits an event if the user expands or collapses the sidebar
-         */
-        "onOpenChange"?: (event: InoSidebarCustomEvent<any>) => void;
-        /**
-          * Expands the sidebar
-         */
-        "open"?: boolean;
-    }
     interface InoSnackbar {
         /**
           * The text to display for the action button. If no text is defined, the snack bar is displayed in an alternative feedback style.
@@ -3058,10 +2972,6 @@ declare namespace LocalJSX {
     }
     interface InoSpinner {
         /**
-          * The name of the color scheme which is used to style the background and outline of this component. Possible values: `primary` (default),  `secondary`, `success`, `warning`, `error`, `light`, `dark`.
-         */
-        "colorScheme"?: ColorScheme;
-        /**
           * The height of this spinner (default = 40)
          */
         "height"?: number;
@@ -3083,10 +2993,6 @@ declare namespace LocalJSX {
           * Marks this element as checked. (**unmanaged**)
          */
         "checked"?: boolean;
-        /**
-          * The name of the color scheme which is used to style the background and outline of this component. Possible values: `primary` (default),  `secondary`, `success`, `warning`, `error`, `light`, `dark`.
-         */
-        "colorScheme"?: ColorScheme;
         /**
           * Disables this element.
          */
@@ -3278,9 +3184,9 @@ declare namespace LocalJSX {
     }
     interface InoTooltip {
         /**
-          * Sets the color scheme of the tooltip Valid options include: `primary`, `secondary` `light`, `transparent`
+          * Sets the color scheme of the tooltip Valid options include: `primary`, `transparent`
          */
-        "colorScheme"?: 'primary' | 'secondary' | 'light' | 'transparent';
+        "colorScheme"?: 'primary' | 'transparent';
         /**
           * The target id the tooltip belongs to. If not given, the tooltip is attached to the parent component.
          */
@@ -3313,7 +3219,6 @@ declare namespace LocalJSX {
         "ino-dialog": InoDialog;
         "ino-fab": InoFab;
         "ino-fab-set": InoFabSet;
-        "ino-form-row": InoFormRow;
         "ino-header": InoHeader;
         "ino-icon": InoIcon;
         "ino-icon-button": InoIconButton;
@@ -3339,7 +3244,6 @@ declare namespace LocalJSX {
         "ino-segment-button": InoSegmentButton;
         "ino-segment-group": InoSegmentGroup;
         "ino-select": InoSelect;
-        "ino-sidebar": InoSidebar;
         "ino-snackbar": InoSnackbar;
         "ino-spinner": InoSpinner;
         "ino-switch": InoSwitch;
@@ -3368,7 +3272,6 @@ declare module "@stencil/core" {
             "ino-dialog": LocalJSX.InoDialog & JSXBase.HTMLAttributes<HTMLInoDialogElement>;
             "ino-fab": LocalJSX.InoFab & JSXBase.HTMLAttributes<HTMLInoFabElement>;
             "ino-fab-set": LocalJSX.InoFabSet & JSXBase.HTMLAttributes<HTMLInoFabSetElement>;
-            "ino-form-row": LocalJSX.InoFormRow & JSXBase.HTMLAttributes<HTMLInoFormRowElement>;
             "ino-header": LocalJSX.InoHeader & JSXBase.HTMLAttributes<HTMLInoHeaderElement>;
             "ino-icon": LocalJSX.InoIcon & JSXBase.HTMLAttributes<HTMLInoIconElement>;
             "ino-icon-button": LocalJSX.InoIconButton & JSXBase.HTMLAttributes<HTMLInoIconButtonElement>;
@@ -3394,7 +3297,6 @@ declare module "@stencil/core" {
             "ino-segment-button": LocalJSX.InoSegmentButton & JSXBase.HTMLAttributes<HTMLInoSegmentButtonElement>;
             "ino-segment-group": LocalJSX.InoSegmentGroup & JSXBase.HTMLAttributes<HTMLInoSegmentGroupElement>;
             "ino-select": LocalJSX.InoSelect & JSXBase.HTMLAttributes<HTMLInoSelectElement>;
-            "ino-sidebar": LocalJSX.InoSidebar & JSXBase.HTMLAttributes<HTMLInoSidebarElement>;
             "ino-snackbar": LocalJSX.InoSnackbar & JSXBase.HTMLAttributes<HTMLInoSnackbarElement>;
             "ino-spinner": LocalJSX.InoSpinner & JSXBase.HTMLAttributes<HTMLInoSpinnerElement>;
             "ino-switch": LocalJSX.InoSwitch & JSXBase.HTMLAttributes<HTMLInoSwitchElement>;