+++ packages/elements/src/components/ino-carousel/ino-carousel.e2e.ts
@@ -39,7 +39,6 @@ describe('ino-carousel', () => {
 describe('InoCarousel', () => {
   let page: E2EPage;
   let inoCarousel: E2EElement;
-  let inoCarouselWrapper: E2EElement;
   let iconArrowRight: E2EElement;
   let iconArrowLeft: E2EElement;
 
@@ -53,7 +52,6 @@ describe('InoCarousel', () => {
     </ino-carousel>
     `);
     inoCarousel = await page.find('ino-carousel');
-    inoCarouselWrapper = await page.find('ino-carousel > div');
     iconArrowLeft = await page.find('.ino-carousel__left-arrow > ino-icon-button')
     iconArrowRight = await page.find('.ino-carousel__right-arrow > ino-icon-button')
   });
@@ -87,7 +85,7 @@ describe('InoCarousel', () => {
   it('should show last slide in carousel upon clicking left arrow icon if first slide is showing', async () => {
     inoCarousel.setAttribute('value', 'a');
     await page.waitForChanges()
-    
+
     const eventSpy = await page.spyOnEvent('valueChange');
     await simulateIconClick('Left');
     expect(eventSpy).toHaveReceivedEventDetail('c');