+++ packages/elements/src/components/ino-tooltip/ino-tooltip.tsx
@@ -28,9 +28,9 @@ export class Tooltip implements ComponentInterface {
 
   /**
    * Sets the color scheme of the tooltip
-   * Valid options include: `primary`, `secondary` `light`, `transparent`
+   * Valid options include: `primary`, `transparent`
    */
-  @Prop() colorScheme: 'primary' | 'secondary' | 'light' | 'transparent' =
+  @Prop() colorScheme: 'primary' | 'transparent' =
     'primary';
 
   /**