+++ packages/elements/stencil.config.ts
@@ -1,8 +1,8 @@
 import { Config } from '@stencil/core';
 import { sass } from '@stencil/sass';
 import { angularOutputTarget } from '@stencil/angular-output-target';
-import { vueOutputTarget } from './vue-output-target';
-import { JsonDocsOutputTarget } from './json-docs-output-target';
+import { vueOutputTarget } from './output-targets/vue-output-target';
+import { JsonDocsOutputTarget } from './output-targets/json-docs-output-target';
 import { reactOutputTarget as react } from '@stencil/react-output-target';
 
 const angularDirectivesPath = '../elements-angular/elements/src/directives';
@@ -22,21 +22,22 @@ export const config: Config = {
     experimentalImportInjection: true,
   },
   globalScript: './src/util/import-fonts.ts',
+  globalStyle: './src/global/styles.scss',
   enableCache: true,
   sourceMap: process.env.NODE_ENV === 'development',
   namespace: 'inovex-elements',
   outputTargets: [
-    react({
-      componentCorePackage: '@inovex.de/elements',
-      proxiesFile: `${reactProxyPath}/index.ts`,
-      includeDefineCustomElements: true,
-    }),
     {
       type: 'dist',
       copy: [{ src: 'assets/ino-icon', dest: 'ino-icon' }],
     },
     { type: 'docs-readme' },
     JsonDocsOutputTarget,
+    react({
+      componentCorePackage: '@inovex.de/elements',
+      proxiesFile: `${reactProxyPath}/index.ts`,
+      includeDefineCustomElements: true,
+    }),
     angularOutputTarget({
       componentCorePackage: '@inovex.de/elements',
       directivesProxyFile: `${angularDirectivesPath}/proxies.ts`,
@@ -70,6 +71,7 @@ export const config: Config = {
             'ino-input',
             'ino-radio-group',
             'ino-range',
+            'ino-segment-group',
             'ino-select',
             'ino-textarea',
           ],