+++ packages/storybook/config/manager.js
@@ -5,4 +5,8 @@ import theme from './theme';
 addons.setConfig({
   panelPosition: 'bottom',
   theme,
+  sidebar: {
+    showRoots: true,
+    collapsedRoots: ['input', 'buttons', 'structure', 'graphic', 'notification'],
+  },
 });