+++ packages/storybook/src/stories/ino-switch/ino-switch.stories.ts
@@ -2,7 +2,7 @@ import { Components } from '@inovex.de/elements';
 import { Meta } from '@storybook/web-components';
 import { html } from 'lit-html';
 import { TemplateGenerator } from '../template-generator';
-import { decorateStoryWithClass } from '../utils';
+import { decorateStoryWithClass, withIconControl } from '../utils';
 import './ino-switch.scss';
 
 const eventHandler = (e: CustomEvent<boolean>) =>
@@ -11,9 +11,8 @@ const eventHandler = (e: CustomEvent<boolean>) =>
 export default {
   title: 'Input/ino-switch',
   component: 'ino-switch',
-  decorators: [(story) => decorateStoryWithClass(story, 'story-switch')],
+  decorators: [story => decorateStoryWithClass(story, 'story-switch')],
   args: {
-    colorScheme: 'primary',
     checked: false,
     disabled: false,
     name: '',
@@ -22,10 +21,9 @@ export default {
 
 const template = new TemplateGenerator<Components.InoSwitch>(
   'ino-switch',
-  (args) => html`
+  args => html`
     <ino-switch
       checked="${args.checked}"
-      color-scheme="${args.colorScheme}"
       disabled="${args.disabled}"
       name="${args.name}"
       @checkedChange="${eventHandler}"
@@ -38,3 +36,26 @@ const template = new TemplateGenerator<Components.InoSwitch>(
 export const Playground = template.generatePlaygroundStory();
 export const Checked = template.generateStoryForProp('checked', true);
 export const Disabled = template.generateStoryForProp('disabled', true);
+
+const templateIconsSlots = new TemplateGenerator<Components.InoSwitch>(
+  'ino-switch',
+  args => html`
+    <ino-switch
+      checked="${args.checked}"
+      disabled="${args.disabled}"
+      name="${args.name}"
+      @checkedChange="${eventHandler}"
+    >
+      <ino-icon slot="leading" icon="offboarding"></ino-icon>
+      Switch Label
+      <ino-icon slot="trailing" icon="onboarding"></ino-icon>
+    </ino-switch>
+  `
+);
+
+/**
+ * Can be used to toggle between to state disabled by two icons.
+ * Use the `leading` and `trailing` slots.
+ * Works best with `<ino-icon>`.
+ */
+export const ToggleIcon = templateIconsSlots.generatePlaygroundStory();